save call slots

+++ contracts/GateKeeper.sol
@@ -5,19 +5,19 @@ import "@openzeppelin/contracts/token/ERC1155/ERC1155.sol";
 
 
 contract GateKeeper is ERC1155 {
-    mapping(uint256 => address) public roomOwners;
+    mapping(uint128 => address) public roomOwners;
 
-    mapping(uint256 => int256) public roomIds;
-    mapping(int256 => uint256) public IdsOfRooms;
+    mapping(uint128 => int128) public roomIds;
+    mapping(int128 => uint128) public IdsOfRooms;
 
-    mapping(address => int256) public idOfUsers;
-    mapping(int256 => address) public addressOfUsers;
+    mapping(address => int128) public idOfUsers;
+    mapping(int128 => address) public addressOfUsers;
 
-    uint256 public nbRooms = 0;
+    uint128 public nbRooms = 0;
 
     constructor() ERC1155("https://ipfs.io/ipfs/QmWXJXRdExse2YHRY21Wvh4pjRxNRQcWVhcKw4DLVnqGqs/") {}
 
-    function createRoom(int256 _roomId, uint256 _supply) public {
+    function createRoom(int128 _roomId, uint128 _supply) public {
         require(IdsOfRooms[_roomId] == 0);
         nbRooms++;
         roomOwners[nbRooms] = msg.sender;
@@ -26,17 +26,17 @@ contract GateKeeper is ERC1155 {
         _mint(msg.sender, nbRooms, _supply, "");
     }
 
-    function mint(uint256 _roomId, uint256 _quantity) public {
+    function mint(uint128 _roomId, uint128 _quantity) public {
         require(msg.sender == roomOwners[_roomId], "Not owner");
         _mint(msg.sender, nbRooms, _quantity, "");
     }
 
-    function transferRoomOwnership(uint256 _roomId, address _newOwner) public {
+    function transferRoomOwnership(uint128 _roomId, address _newOwner) public {
         require(msg.sender == roomOwners[_roomId], "Not owner");
         roomOwners[_roomId] = _newOwner;
     }
 
-    function register(int256 _userId) public {
+    function register(int128 _userId) public {
         require(addressOfUsers[_userId] == address(0), "Already taken");
         addressOfUsers[_userId] = msg.sender;
         idOfUsers[msg.sender] = _userId;
@@ -47,7 +47,7 @@ contract GateKeeper is ERC1155 {
         idOfUsers[msg.sender] = 0;
     }
 
-    function isGateOpen(int256 _userId, int256 _chatId) public view returns (bool) {
+    function isGateOpen(int128 _userId, int128 _chatId) public view returns (bool) {
         return (this.balanceOf(addressOfUsers[_userId], IdsOfRooms[_chatId]) > 0 && addressOfUsers[_userId] != address(0));
     }
